QemuHooks
